started basic tetra implemenation moved all tetra code to TetraMeshSpatialGrid removed unnecessary code/classes and fixed small oversights added tetgen LICENSE
